Allow web to initiate the connection.
We now defer the connection of the application to the robot serial
interface until such time as the web interface requests it. The primary
reason for this is that when the Pi starts up, it is unlikely that the
robot has been turned on and is ready for connection. This is an issue
we had with telep 1.